requirements.txt
Django==3.0.3
celery==4.4.1
django-celery-beat==2.3.0
redis==4.3.4
django-tinymce==2.8.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-cors-headers==3.2.1
djangorestframework-simplejwt==4.5.0
django-filter==2.2.0
Pillow==9.2.0